+++ cloud-controller-manager/cmd/digitalocean-cloud-controller-manager/main.go
@@ -27,7 +27,8 @@ import (
 	_ "k8s.io/kubernetes/pkg/client/metrics/prometheus" // for client metric registration
 	_ "k8s.io/kubernetes/pkg/version/prometheus"        // for version metric registration
 
-	_ "github.com/digitalocean/digitalocean-cloud-controller-manager/cloud-controller-manager/do"
+	"github.com/digitalocean/digitalocean-cloud-controller-manager/cloud-controller-manager/do"
+	"github.com/spf13/pflag"
 )
 
 func init() {
@@ -41,6 +42,34 @@ func main() {
 	flag.CommandLine.String("cloud-provider-gce-lb-src-cidrs", "", "NOT USED (workaround for https://github.com/kubernetes/kubernetes/issues/76205)")
 
 	command := app.NewCloudControllerManagerCommand()
+
+	// Set static flags for which we know the values.
+	command.Flags().VisitAll(func(fl *pflag.Flag) {
+		var err error
+		switch fl.Name {
+		case "allow-untagged-cloud",
+			// Untagged clouds must be enabled explicitly as they were once marked
+			// deprecated. See
+			// https://github.com/kubernetes/cloud-provider/issues/12 for an ongoing
+			// discussion on whether that is to be changed or not.
+			"authentication-skip-lookup":
+			// Prevent reaching out to an authentication-related ConfigMap that
+			// we do not need, and thus do not intend to create RBAC permissions
+			// for. See also
+			// https://github.com/digitalocean/digitalocean-cloud-controller-manager/issues/217
+			// and https://github.com/kubernetes/cloud-provider/issues/29.
+			err = fl.Value.Set("true")
+		case "cloud-provider":
+			// Specify the name we register our own cloud provider implementation
+			// for.
+			err = fl.Value.Set(do.ProviderName)
+		}
+
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "failed to set flag %q: %s\n", fl.Name, err)
+			os.Exit(1)
+		}
+	})
 
 	// (The following comment is copied from upstream:)
 	// TODO: once we switch everything over to Cobra commands, we can go back to calling